+++ R/cellAggregation.R
@@ -44,42 +44,50 @@
 aggregateAcrossCells <- function(x, factors, num.threads = 1) {
   f <- lapply(factors, factor)
   f0 <- lapply(f, as.integer)
-
+  
   x <- beachmat::initializeCpp(x)
   output <- aggregate_across_cells(x, f0, num.threads)
   rownames(output$sums) <- rownames(output$detected) <- rownames(x)
-
+  
   for (i in seq_along(output$combinations)) {
     current <- output$combinations[[i]]
     output$combinations[[i]] <- levels(f[[i]])[current]
   }
-
+  
   names(output$combinations) <- names(factors)
   output$combinations <- data.frame(output$combinations)
   output$index <- output$index + 1L
-
+  
   output
 }
 
 
 #' @importFrom SummarizedExperiment colData<-
 #' @importFrom SingleCellExperiment applySCE
-.aggregateCells <- function(cellNum, expMatrix, peakMatrix, caller_env, useDim,
-                            exp_assay, peak_assay, BPPARAM, clusters=NULL){
+.aggregateCells <- function(cellNum, 
+                            expMatrix, 
+                            peakMatrix, 
+                            caller_env, 
+                            useDim,
+                            exp_assay, 
+                            peak_assay, 
+                            BPPARAM, 
+                            clusters=NULL){
   message("performing pseudobulk using an average of ", cellNum, " cells")
-
+  
   if (!is.null(clusters)) {
     kclusters <- list()
     barcodes <- list()
-
+    
     #add cluster info to expMatrix
     colData(expMatrix)[, "cluster_for_pseudobulk"] <- clusters
-
+    
     # K-means clustering
     for (cluster in unique(clusters)) {
       sce <- expMatrix[, which(clusters == cluster)]
       kNum <- trunc(ncol(sce)/cellNum)
-      kclusters[[cluster]] <- scran::clusterCells(sce, use.dimred = useDim,
+      kclusters[[cluster]] <- scran::clusterCells(sce, 
+                                                  use.dimred = useDim,
                                                   BLUSPARAM = bluster::KmeansParam(centers = kNum, iter.max = 5000))
       barcodes[[cluster]] <- names(kclusters[[cluster]])
       kclusters[[cluster]] <- paste(cluster, kclusters[[cluster]], sep = "_")
@@ -87,24 +95,31 @@ aggregateAcrossCells <- function(x, factors, num.threads = 1) {
     kclusters <- unlist(kclusters)
     barcodes <- unlist(barcodes)
     names(kclusters) <- barcodes
-
-
+    
+    
   } else {
-    kclusters <- scran::clusterCells(expMatrix, use.dimred = useDim, BLUSPARAM = bluster::KmeansParam(centers = trunc(ncol(peakMatrix)/cellNum),
-                                                                                                      iter.max = 5000))
+    kclusters <- scran::clusterCells(expMatrix, 
+                                     use.dimred = useDim, 
+                                     BLUSPARAM = bluster::KmeansParam(centers = trunc(ncol(peakMatrix)/cellNum), iter.max = 5000))
   }
-
+  
   kclusters <- kclusters[colnames(expMatrix)]
-
+  
   #replace clusters with clusters of pseudobulked samples
-
-  expMatrix <- aggregateAcrossCellsFast(expMatrix, clusters = kclusters, fun_name="sum",
-                                         assay.name = exp_assay)
-  peakMatrix <- aggregateAcrossCellsFast(peakMatrix, clusters = kclusters, fun_name="sum",
-                                          assay.name = peak_assay)
-
-  if (!is.null(clusters))
+  
+  expMatrix <- aggregateAcrossCellsFast(expMatrix, 
+                                        clusters = kclusters, 
+                                        fun_name="sum",
+                                        assay.name = exp_assay)
+  peakMatrix <- aggregateAcrossCellsFast(peakMatrix, 
+                                         clusters = kclusters, 
+                                         fun_name="sum",
+                                         assay.name = peak_assay)
+  
+  if (!is.null(clusters)) {
     clusters <- colData(expMatrix)[, "cluster_for_pseudobulk"]
+  }
+  
   assign("expMatrix",expMatrix,envir = caller_env)
   assign("peakMatrix",peakMatrix,envir = caller_env)
   assign("clusters",clusters,envir = caller_env)
@@ -118,7 +133,7 @@ aggregateAcrossCells <- function(x, factors, num.threads = 1) {
 #' for each cluster/sample combination. It is wrapped around `aggregateAcrossCells`,
 #' which relies on the C++ code.
 #'
-#' @param sce A SingleCellExperiment object
+#' @param sce A SingleCellExperiment, SummarizedExperiment or RangedSummarizedExperiment object
 #' @param clusters A vector used as a grouping variable. The length should be equal to
 #' the number of cells.
 #' @param assay.name A character indicating the name of the assay containing the
@@ -141,62 +156,99 @@ aggregateAcrossCells <- function(x, factors, num.threads = 1) {
 #' ids <- sample(LETTERS[1:5], ncol(example_sce), replace=TRUE)
 #' out <- aggregateAcrossCellsFast(example_sce, ids)
 #' @export
-aggregateAcrossCellsFast <- function(sce, clusters, assay.name="counts", fun_name=c("mean", "sum"),
-                                     num.threads=1, aggregateColData = TRUE) {
-  .validate_input_sce(expMatrix=sce, exp_assay=assay.name, env=environment(), show_warning = FALSE)
+aggregateAcrossCellsFast <- function(sce, 
+                                     clusters, 
+                                     assay.name="counts", 
+                                     fun_name=c("mean", "sum"),
+                                     num.threads=1, 
+                                     aggregateColData = TRUE) {
+  # validate inputs
+  .validate_input_sce(SCE=sce, 
+                      assay_name=assay.name,
+                      accepted_classes = c("SingleCellExperiment", "SummarizedExperiment", "RangedSummarizedExperiment"))
   .validate_clusters(clusters, sce)
+  clusters <- as.vector(clusters)
   fun_name <- match.arg(fun_name, several.ok = FALSE)
+  
   # aggregate counts in assay
-  if(!is.null(assay.name)) x <- setNames(assays(sce)[assay.name], assay.name)
-  else x <- setNames(assays(sce), names(assays(sce)))
+  if(!is.null(assay.name)) {
+    x <- setNames(assays(sce)[assay.name], assay.name)
+  } else {
+    x <- setNames(assays(sce), names(assays(sce)))
+  }
+  
   aggr.counts <- lapply(x, aggregateAcrossCells, factors = list(clusters), num.threads=num.threads)
-  if(fun_name=="sum") assay_matrices <- setNames(lapply(aggr.counts, "[[", "sums"), names(x))
-  else assay_matrices <- setNames(lapply(aggr.counts, function(x) t(t(x$sums)/x$counts)), names(x)) #mean
-  altExps_list <- NULL
-  if(length(altExps(sce))>0){
-    altExps_list <- lapply(altExps(sce), aggregateAcrossCellsFast, clusters, NULL, fun_name,
-                           FALSE)
+  if(fun_name=="sum") {
+    assay_matrices <- setNames(lapply(aggr.counts, "[[", "sums"), names(x))
+  } else {
+    assay_matrices <- setNames(lapply(aggr.counts, function(x) t(t(x$sums)/x$counts)), names(x)) #mean
+    altExps_list <- NULL
+  }
+  
+  if(is(sce, "SingleCellExperiment") && length(altExps(sce))>0){
+    altExps_list <- lapply(altExps(sce), aggregateAcrossCellsFast, clusters, NULL, fun_name, FALSE)
     names(altExps_list) <- altExpNames(sce)
   }
-
+  
   # reassemble the singleCellExperiment object
   sce.bulk <- SingleCellExperiment(assay_matrices,
                                    rowData = rowData(sce))
   rownames(colData(sce.bulk)) <- colData(sce.bulk)$idx <- aggr.counts[[1]]$combinations[,1]
   colData(sce.bulk)$ncells <- aggr.counts[[1]]$counts
-  altExps(sce.bulk) <- altExps_list
+  if(is(sce, "SingleCellExperiment") && length(altExps(sce))>0){
+    altExps(sce.bulk) <- altExps_list
+  }
+  
   if(aggregateColData){
     colData.sce.consistent <- .select_consistent_columns(colData(sce), clusters)
-    first.position <- match(aggr.counts[[1]]$combinations[,1],clusters)
-    if(!is.null(colData.sce.consistent))
-      colData(sce.bulk) <- cbind(colData(sce.bulk), colData.sce.consistent[first.position,, drop=FALSE])
+    if(!is.null(colData.sce.consistent)){
+      duplicated_colnames <- intersect(colnames(colData(sce.bulk)), colnames(colData.sce.consistent))
+      if(length(duplicated_colnames)>0){
+        stop(spritnf("The following columns are already present in the colData: %s", paste(duplicated_colnames,collapse=", ")))
+      }
+      # one row per cluster
+      unique_clusters_idx <- match(aggr.counts[[1]]$combinations[,1], clusters)
+      colData(sce.bulk) <- cbind(colData(sce.bulk), colData.sce.consistent[unique_clusters_idx,,drop=FALSE])
+    }
   }
   rowRanges(sce.bulk) <- rowRanges(sce)
   sce.bulk
 }
 
 .select_consistent_columns <- function(df, ids){
-  if(ncol(df)==1){
-    if(is.numeric(ncol(df[,1]))) df[[1]] <- .select_consistent_columns(df[,1],ids)
-    else df <- df[,.is_consistent(df[,1], ids),drop=FALSE]
+  if(is(df,"DataFrame")){ # handle DataFrame separately to preserve its hierarchical structure
+    current_col = 0
+    for (i in seq_len(ncol(df))){
+      current_col = current_col+1
+      if(length(dim(df[,current_col]))<2) { # select vectors and 1-dim arrays
+        if(!.is_consistent(df[,current_col],ids)) {
+          df[[current_col]] <- NULL
+          current_col = current_col-1
+        }
+      } else{
+        consistent_data <- .select_consistent_columns(df[,current_col], ids)
+        df[[current_col]] <- consistent_data
+        if(is.null(consistent_data)){
+          current_col = current_col-1
+        }
+      }
+    }
+  } else{
+    preserve_columns <- integer(0)
+    for(i in seq_len(ncol(df))){
+      if(.is_consistent(df[,i],ids)){
+        preserve_columns <- c(preserve_columns, i)
+      }
+    }
+    df <- df[, preserve_columns, drop=FALSE]
   }
-  else{
-    cols <- lapply(seq_len(ncol(df)), .return_consistent_column, df, ids)
-    # remove NULLs and empty objects and bind columns into the final object
-    df <- do.call(cbind, cols[unlist(lapply(cols, function(x) length(x)!=0))])
+  if(ncol(df)==0){
+    df <- NULL
   }
-  if(is.null(df) || ncol(df)==0) return(NULL) # otherwise some columns might be composed of a DataFrame with one empty column
   df
-}
-
-.return_consistent_column <- function(i, df, ids){
-  if(length(dim(df[,i]))<2) {        # select vectors and 1-dim arrays
-    if(.is_consistent(df[,i],ids)) return(df[,i,drop=FALSE])
-    else return(NULL)
-  }
-  else .select_consistent_columns(df[,i,drop=FALSE], ids)
 }
 
 .is_consistent <- function(x1, x2){
   all(unlist(lapply(split(x1, x2), function(x) length(unique(x))==1)))
 }
+