+++ R/epiregulon-package.R
@@ -3,5 +3,7 @@
 
 ## usethis namespace: start
 #' @importFrom lifecycle deprecated
+#' @useDynLib epiregulon
+#' @importFrom Rcpp sourceCpp
 ## usethis namespace: end
 NULL