+++ R/validation.R
@@ -14,51 +14,58 @@
 .balance_check <- function(peak_cutoff, exp_cutoff, peakMatrix, expMatrix){
   if (!is.null(peak_cutoff)) {
     prop <- sum(peakMatrix > peak_cutoff)/prod(dim(peakMatrix))
-    if (prop < 1e-04 | prop > 0.9999)
+    if (prop < 1e-04 | prop > 0.9999) {
       warning(sprintf("Strong inbalance between groups after applying cutoff to peakMatrix. Consider %s value of the peak_cutoff",
                       c("increasing", "decreasing")[(prop < 1e-04) + 1]))
+    }
   }
   if (!is.null(exp_cutoff)) {
     prop <- sum(expMatrix > exp_cutoff)/prod(dim(expMatrix))
-    if (prop < 1e-04 | prop > 0.9999)
+    if (prop < 1e-04 | prop > 0.9999) {
       warning(sprintf("Strong inbalance between groups after applying cutoff to expMatrix. Consider %s value of the exp_cutoff",
                       c("increasing", "decreasing")[(prop < 1e-04) + 1]))
+    }
   }
 }
-#' @importFrom SummarizedExperiment assay<-
 
-.validate_input_sce <- function(expMatrix, exp_assay, peakMatrix=NULL, peak_assay=NULL, tf_re.merge=FALSE, env, row.ranges=FALSE,
-                                show_warning = TRUE){
-    checkmate::assert_class(expMatrix, "SingleCellExperiment")
-    stopifnot(exp_assay %in% names(assays(expMatrix)))
-    if (any(dim(expMatrix) == 0)) stop("SingleCellExperiment with no data")
-    conversion_failed <- FALSE
-    tryCatch(assay(expMatrix, exp_assay) <- as(assay(expMatrix, exp_assay), "CsparseMatrix"),
-                                          error = function(cond) {
-                                              if(show_warning) message("Gene expression assay cannot be coerced to CSparseMatrix. This might affect negatively function performance.")
-                                              conversion_failed <<- TRUE}
-                                          )
-    if(!conversion_failed) assign("expMatrix", expMatrix, envir = env)
-    if(tf_re.merge | !is.null(peakMatrix)){
-        checkmate::assert_class(peakMatrix, "SingleCellExperiment")
-        stopifnot(peak_assay %in% names(assays(peakMatrix)))
-        stopifnot(ncol(peakMatrix) == ncol(expMatrix))
-        if(nrow(peakMatrix)==0) stop("peakMatrix with no data")
-        conversion_failed <- FALSE
-        tryCatch(assay(peakMatrix, peak_assay) <- as(assay(peakMatrix, peak_assay), "CsparseMatrix"),
-                                                error = function(cond) {
-                                                    if(show_warning) message("Peak count assay cannot be coerced to CSparseMatrix. This might affect negatively function performance.")
-                                                    conversion_failed <<- TRUE}
-                                                )
-        if(!conversion_failed) assign("peakMatrix", peakMatrix, envir = env)
+
+.validate_input_sce <- function(SCE,
+                                assay_name,
+                                row.ranges=FALSE,
+                                accepted_classes = c("SingleCellExperiment", "RangedSummarizedExperiment"),
+                                unique_features = FALSE){
+    checkmate::assert_multi_class(SCE, accepted_classes)
+    stopifnot(assay_name %in% names(assays(SCE)))
+    data_object_name <- as.character(substitute(SCE))
+    if (any(dim(SCE) == 0)){
+        stop(sprintf("%s with no data", data_object_name))
+    }
+    if (unique_features){
+        if (any(duplicated(rownames(SCE)))){
+            stop(sprintf("Feature names in %s should be unique", data_object_name))
+        }
     }
     if(row.ranges){
-        if (length(rowRanges(peakMatrix)) == 0) {
-            stop("peakMatrix should contain non-empty rowRanges")
+        if (length(rowRanges(SCE)) == 0) {
+            stop(sprintf("%s should contain non-empty rowRanges", data_object_name))
         }
-        checkmate::assert_class(rowRanges(peakMatrix), "GRanges")
-        if (length(rowRanges(expMatrix)) == 0) {
-            stop("expMatrix should contain non-empty rowRanges")
+        checkmate::assert_class(rowRanges(SCE), "GRanges")
+        if (length(rowRanges(SCE)) == 0) {
+            stop(sprintf("%s should contain non-empty rowRanges", data_object_name))
         }
+    }
+}
+
+.validate_regulon <- function(regulon, required_columns = c("tf", "target", "idxATAC")){
+    checkmate::assert_multi_class(regulon, c("DataFrame", "data.frame", "DFrame"))
+    if (!all(required_columns %in% colnames(regulon))) {
+        stop(paste("regulon should contain the following columns: ", paste(required_columns, sep = ", "),sep = ""))
+    }
+    if (nrow(regulon)==0){
+        stop("regulon should not be empty")
+    }
+    columns_with_NA <- required_columns[unlist(lapply(regulon[,required_columns], function(x) any(is.na(x))))]
+    if (length(columns_with_NA)>0){
+        warning(paste("The following regulon column(s) contain NA value(s): ", paste(columns_with_NA, sep = ", "),sep = ""))
     }
 }