+++ tests/testthat/test-calculateActivity.R
@@ -1,9 +1,9 @@
 regulon <- S4Vectors::DataFrame(tf = rep(LETTERS[1:5], times = 1:5),
                       target = LETTERS [6:20],
-                      weight = seq(0,1,length.out = 15))
+                      weight = seq(0,1,length.out = 15),
+                      idxATAC=0)
 
-# account for scenario when weight==NA
-regulon$weight[2] <- NA
+
 
 geneExpressionMatrix <- matrix(0, nrow = 15, ncol = 100, dimnames = list(LETTERS[6:20], NULL))
 geneExpressionMatrix[1:(88*17)] <- rep(c(0, 0, 1, 2, 3, 0, 0, 1, 2, 0, 2, 0, 0, 0, 1, 3, 0), 88)
@@ -17,9 +17,6 @@ tf_tg_matrix <- as(tf_tg_matrix, "sparseMatrix")
 for (i in seq_len(nrow(regulon))){
   tf_tg_matrix[regulon[i,"target"], regulon[i,"tf"]] <- regulon[i, "weight"]
 }
-
-# replace NA with 0 in tf_tg_matrix
-tf_tg_matrix[is.na(tf_tg_matrix)] <- 0
 
 tf_tg_matrix2 <- createTfTgMat(regulon, "weight", clusters=NULL)
 
@@ -39,7 +36,7 @@ test_that("calculateScore works correctly", {
 
 ### test calculateFrequency
 # divide by the number of target genes
-freq <- c(1,3,4,5)
+freq <- c(2,3,4,5)
 names(freq) <- c("B","C","D", "E") #no weight != 0 in A, 1 NA weight changed to 0 in B
 freq2 <- calculateFrequency(regulon=regulon, mode="weight")
 freq3 <- as(freq2, "double")
@@ -50,7 +47,7 @@ test_that("calculateFrequency works correctly", {
 })
 
 ### test normalizeByFrequency
-activity_matrix.norm <- activity_matrix/c(1,1,3,4,5)
+activity_matrix.norm <- activity_matrix/c(1,2,3,4,5)
 activity_matrix.norm2 <- normalizeByFrequency(activity_matrix2, freq2)
 test_that("normalizeByFrequency works correctly", {
   expect_equal( as.matrix(Matrix::t(activity_matrix.norm2)), activity_matrix.norm)
@@ -58,6 +55,7 @@ test_that("normalizeByFrequency works correctly", {
 
 ### test calculateActivity
 sce <- SingleCellExperiment(assay = list(counts = geneExpressionMatrix))
+
 activity_matrix.3 <- calculateActivity(sce, regulon = regulon, exp_assay = "counts")
 
 test_that("calculateActivity works correctly", {
@@ -66,20 +64,34 @@ test_that("calculateActivity works correctly", {
                )
 })
 
+# account for scenario when weight==NA
+substituted_weight <- regulon$weight[2]
+regulon$weight[2] <- NA
+
+test_that("calculateActivity warns when NA is present", {
+    expect_warning(calculateActivity(sce, regulon = regulon, exp_assay = "counts"))
+})
+
+# restore original value
+regulon$weight[2] <- substituted_weight
+
+# drop first column since all corresponding weights are zero
+tf_tg_matrix <- tf_tg_matrix[,-1]
+
 ### subtract mean gene expression (centering at zero)
 geneExpressionMatrix.mean <- sweep(geneExpressionMatrix, 1,
                               Matrix::rowMeans(geneExpressionMatrix),"-")
 activity_matrix.center <- as.matrix(Matrix::t(Matrix::t(geneExpressionMatrix.mean)
                                               %*% tf_tg_matrix))
 
-activity_matrix.center.norm <- activity_matrix.center/c(1,1,3,4,5)
+activity_matrix.center.norm <- activity_matrix.center/c(2,3,4,5)
 
 test_that("calculateActivity normalizes correctly (zero-centering)", {
   expect_equal(as.matrix(calculateActivity(sce,
                                            regulon = regulon,
                                            exp_assay = "counts",
                                            normalize = TRUE)),
-               activity_matrix.center.norm[Matrix:::rowSums(activity_matrix.center.norm)!=0,])
+               activity_matrix.center.norm)
 })
 
 ###################################################################################
@@ -143,7 +155,7 @@ freq3 <- freq4 <- initiateMatCluster(clusters, nrow = length(unique(regulon$tf))
 rownames(freq3) <- rownames(freq4) <- unique(regulon$tf)
 freq3[,"C1"] <- c(1,2,3,4,4)
 freq3[,"C2"] <- 1:5
-freq3[,"all"] <- c(1,1,3,4,5)
+freq3[,"all"] <- c(1,2,3,4,5)
 
 freq6 <- calculateFrequency(freq4, regulon, mode="weight")
 
@@ -171,7 +183,7 @@ activity_matrix.norm5 <- calculateActivity(sce,
                                            FUN = "mean")
 test_that("calculateActivity works correctly with clusters", {
   expect_equal(activity_matrix.norm5,
-               activity_matrix.norm3[Matrix:::rowSums(activity_matrix.norm3)!=0,])
+               as.matrix(activity_matrix.norm3[Matrix:::rowSums(activity_matrix.norm3)!=0,]))
 })